Fixed an error which could cause elements to be not selected

+++ src/HLISA/util.py
@@ -16,9 +16,8 @@ class HL_Util:
         y_relative = int(element.rect['y']) - webdriver.execute_script("return window.pageYOffset;")
         counter = 0
         for i in range(10): # Try 10 random positions, as some positions are not in round buttons.
-            x = x_relative + np.random.normal(int(element.rect['width']*0.5), int(element.rect['width']*0.2))
-            y = y_relative + np.random.normal(int(element.rect['height']*0.5), int(element.rect['height']*0.2))
-        
+            x = x_relative + int(np.random.normal(int(element.rect['width']*0.5), int(element.rect['width']*0.2)))
+            y = y_relative + int(np.random.normal(int(element.rect['height']*0.5), int(element.rect['height']*0.2)))
             coords_in_button = webdriver.execute_script("return document.elementFromPoint(" + str(x) + ", " + str(y) + ") === arguments[0];", element)
 
             if coords_in_button:
